==> processing workflow/14_make_final_csv.ipynb <==
import os
import rasterio as rio
import numpy as np
import shapely
import pyproj
import geopandas as gpd
import matplotlib.pyplot as plt
import rioxarray as riox
import rasterio as rio
import xarray as xr
import netCDF4
from osgeo import gdal
import pandas as pd
from datetime import datetime
import dask.array

import sys
sys.path.append('../')
import snowFun

# define folder and file paths
folder_AGVA = os.path.join('C:',os.sep,'Users','lzell','OneDrive - Colostate','Desktop',"AGVA")
folder_dems = os.path.join(folder_AGVA, "DEMs", "time_varying_DEMs", "10m")
folder_class = os.path.join(folder_AGVA, 'classified images', 'S2_Classified_Cloudmasked_Merged')
folder_cloud = os.path.join(folder_AGVA, 'classified images', 'S2_Cloud_Merged')
folder_meta = os.path.join(folder_AGVA, "classified images", "meta csv", "S2")
folder_dhdt = os.path.join(folder_AGVA, "dh_dt")

# open rgi
path_rgi = os.path.join(folder_AGVA, 'RGI', "rgi_2km_o3regions", "rgi_2km_o3regions.shp")
rgi_gdf = gpd.read_file(path_rgi, drop='geometry')

# get list of all the glaciers we have data for
folder_source = os.path.join(folder_AGVA, 'Derived products', 'S2', 'Annual AAs', 'csv')
folder_save = os.path.join(folder_AGVA, 'Derived products', 'S2')

# load rgi names that have been saved to the classified folder
rgis_to_analyze = rgis_to_analyze = list(set( [ i[3:17] for i in os.listdir(os.path.join(folder_save, "Average AAs")) if i[-3:]=='tif' ] ))
rgis_to_analyze.sort()

print(len(rgis_to_analyze))

### Create a single dataframe that hold ELA and AAR info for each year for each glacier

# go through all the saved csvs, make a master csv with rgiid, annual ela/aar info
all_files = [ i for i in os.listdir(os.path.join(folder_source)) if i[-3:]=='csv' ]

all_rgis = []
all_elas_2018 = []
all_elas_2019 = []
all_elas_2020 = []
all_elas_2021 = []
all_elas_2022 = []

all_aars_2018 = []
all_aars_2019 = []
all_aars_2020 = []
all_aars_2021 = []
all_aars_2022 = []

all_og_2018 = []
all_og_2019 = []
all_og_2020 = []
all_og_2021 = []
all_og_2022 = []

all_dates_2018 = []
all_dates_2019 = []
all_dates_2020 = []
all_dates_2021 = []
all_dates_2022 = []

i=0
for f in all_files:
    i+=1
#     if i>300: continue
    
    # open
    df = pd.read_csv(os.path.join(folder_source, f))
    
    # if it doesn't contain aar_infill, add blank column
    if 'aar_infill' not in df.columns:
        df['aar_infill'] = np.nan
    
    # get rgi
    all_rgis.append( f[3:17] )
    
    # add data to lists
    all_elas_2018.append(int(df.loc[0]['ela']))
    all_elas_2019.append(int(df.loc[1]['ela']))
    all_elas_2020.append(int(df.loc[2]['ela']))
    all_elas_2021.append(int(df.loc[3]['ela']))
    all_elas_2022.append(int(df.loc[4]['ela']))
    
    all_aars_2018.append(round(df.loc[0]['aar_infill'],4))
    all_aars_2019.append(round(df.loc[1]['aar_infill'],4))
    all_aars_2020.append(round(df.loc[2]['aar_infill'],4))
    all_aars_2021.append(round(df.loc[3]['aar_infill'],4))
    all_aars_2022.append(round(df.loc[4]['aar_infill'],4))
    
    all_og_2018.append(df.loc[0]['off_glacier'])
    all_og_2019.append(df.loc[1]['off_glacier'])
    all_og_2020.append(df.loc[2]['off_glacier'])
    all_og_2021.append(df.loc[3]['off_glacier'])
    all_og_2022.append(df.loc[4]['off_glacier'])
    
    all_dates_2018.append(df.loc[0]['date'])
    all_dates_2019.append(df.loc[1]['date'])
    all_dates_2020.append(df.loc[2]['date'])
    all_dates_2021.append(df.loc[3]['date'])
    all_dates_2022.append(df.loc[4]['date'])
    
    # get rid of the 'bad' years
#     df = df[df['ela']>0]
    
    # get median ela, aar
#     all_elas.append( np.nanmedian(df['ela']) )
#     all_aars.append( np.nanmedian(df['aar']) )

# save everything
# obs_df = pd.DataFrame({'RGIId':all_rgis, 'ela':all_elas, 'aar':all_aars}) 
obs_df = pd.DataFrame({'RGIId':all_rgis,
                       'ela_2018':all_elas_2018, 'ela_2019':all_elas_2019, 'ela_2020':all_elas_2020, 'ela_2021':all_elas_2021, 'ela_2022':all_elas_2022,
                       'aar_2018':all_aars_2018, 'aar_2019':all_aars_2019, 'aar_2020':all_aars_2020, 'aar_2021':all_aars_2021, 'aar_2022':all_aars_2022,
                       'off_glacier_2018':all_og_2018, 'off_glacier_2019':all_og_2019, 'off_glacier_2020':all_og_2020, 'off_glacier_2021':all_og_2021, 'off_glacier_2022':all_og_2022,
                       'date_2018':all_dates_2018, 'date_2019':all_dates_2019, 'date_2020':all_dates_2020, 'date_2021':all_dates_2021, 'date_2022':all_dates_2022,
                      }) 
print(obs_df.shape)
obs_df.head()

# add rgi info (area, zmin, zmax, etc...) to df
full_data_df = rgi_gdf[rgi_gdf['RGIId'].isin(rgis_to_analyze)]
full_data_df = full_data_df.merge(obs_df, on='RGIId')
print(full_data_df.shape)
full_data_df.head()

# recode the ELA in years where ela is off glacier as Zmax
for y in [2018,2019,2020,2021,2022]:
    full_data_df.loc[full_data_df[f'off_glacier_{y}'].astype(str)=="True", f'ela_{y}'] = full_data_df.loc[full_data_df[f'off_glacier_{y}'].astype(str)=="True", 'Zmax']

# open the ela aar for the 5-year average products
path_open = os.path.join(folder_AGVA, 'Derived products', 'S2', 'Average AAs', 'final_glacier_stats.csv')
average_df = pd.read_csv(path_open)
average_df = average_df.rename({'ela':'ela_avg_product',
                                'aar':'aar_avg_product',
                                'aar_infill':'aar_infill_avg_product',
                                'off_glacier':'off_glacier_avg_product',
                               }, axis=1)
average_df['percent_observed_average_product'] = (average_df['total_observed']/average_df['total_area']).round(4)
to_merge_df = average_df[['RGIId','ela_avg_product','aar_avg_product','aar_infill_avg_product',
                          'off_glacier_avg_product','percent_observed_average_product']]
full_data_df = full_data_df.merge(to_merge_df, on='RGIId', how = 'outer')

full_data_df.loc[full_data_df['off_glacier_avg_product'].astype(str)=="True", 'ela_avg_product'] = full_data_df.loc[full_data_df['off_glacier_avg_product'].astype(str)=="True", 'Zmax']

print(full_data_df.shape)
full_data_df.head()

# add dhdt info to df

# open dhdt
path_dhdt = os.path.join(folder_dhdt, "dh_01_rgi60_pergla_rates.csv")
dhdt_df = pd.read_csv(path_dhdt, usecols=["rgiid","period","dhdt"])

# dropping unwanted rows
dhdt_df = dhdt_df[dhdt_df['rgiid'].isin(rgis_to_analyze)]
dhdt_df = dhdt_df.rename(columns={'rgiid':'RGIId'})

# get only the rows and columns we are interested in (and merge)
dhdt_1520_df = dhdt_df[ dhdt_df['period'].isin(['2015-01-01_2020-01-01']) ].copy().rename(columns={'dhdt':'dhdt_1520'}).drop('period', axis=1)
dhdt_1020_df = dhdt_df[ dhdt_df['period'].isin(['2010-01-01_2020-01-01']) ].copy().rename(columns={'dhdt':'dhdt_1020'}).drop('period', axis=1)
dhdt_df = dhdt_1520_df.merge(dhdt_1020_df, on='RGIId')

print(dhdt_df.shape)
print(dhdt_df.head())

full_dhdt_df = dhdt_df[dhdt_df['RGIId'].isin(rgis_to_analyze)]
full_data_df = full_data_df.merge(full_dhdt_df, on='RGIId')
print(full_data_df.shape)
full_data_df.head()

# add distance from ocean info to df
ocean_df = pd.read_csv(os.path.join(folder_AGVA, "Derived products", "S2", "distance_from_ocean.csv"))
full_data_df = full_data_df.merge(ocean_df, on='RGIId')
print(full_data_df.shape)
full_data_df.head()

# add hypsometric analysis data
hyp_df = pd.read_csv(os.path.join(folder_save, 'Average AAs', 'hyspometric_stats.csv'))
hyp_df = hyp_df.drop(['obs_ela', 'obs_aar'], axis=1)

full_data_df = full_data_df.merge(hyp_df, on='RGIId', how='left')
print(full_data_df.shape)
full_data_df.head()

# add cmip-derived ela prediction and resulting aar
exp_aar_df = pd.read_csv(os.path.join(folder_save, 'Average AAs', 'CMIP_derived_AARs.csv'))
exp_aar_df = exp_aar_df.drop(['hyps_aar'], axis=1).rename({'hyps_aar_predicted':'hyps_aar_cmip'}, axis=1)
# exp_aar_df

# open the glacier-specific future ela changes
path_dela = os.path.join(folder_AGVA, 'Climate', 'CMIP', "rgi_delta_ELA.csv")
dela_df = pd.read_csv(path_dela).rename({'delta_ela':'delta_ela_cmip'}, axis=1)
# dela_df

# merge together
to_add = exp_aar_df.merge(dela_df, on='RGIId', how='left')
full_data_df = full_data_df.merge(to_add, on='RGIId', how='left')
print(full_data_df.shape)
full_data_df.head()

# drop geometry column prior to saving
good_cols = [i for i in full_data_df.columns if i != "geometry"]
full_data_df = full_data_df[good_cols]
full_data_df.head()

out_path = os.path.join(folder_save, 'Final_Data_Table.csv')
full_data_df.to_csv(out_path, index=False)
